:recycle: Divide place table into multiple subcomponents PlaceTableHeaderRow, PlaceTableEmptyMessageRow, PlaceTableRow
